motor speed as float 0-100

+++ mcu/common/include/motors.h
@@ -3,4 +3,4 @@
 int motors_init();
 int is_driving();
 void motors_estop();
-int set_drive(int16_t lhs, int16_t rhs);
+int set_drive(float lhs, float rhs);
